Clean-up animations after no-event timeout
As discussed in issue 1326
(https://github.com/facebook/react/issues/1326) transitionend events are
unreliable; they may not fire because the element is no longer painted,
the browser tab is no longer focused or for a range of other reasons.
This is particularly harmful during leave transitions since the leaving
element will be permanently stuck in the DOM (and possibly visible).
There is already a 5 second timeout in dev mode to warn about the
missing event, this commit uses that timeout to clean-up the animation
(in both dev and non-dev modes).  This is hardly an ideal solution, but
it's much better than the status-quo of failed transition elements
hanging-around indefinitely.

The `if (__DEV__)` wrapper around the `warning` call has been removed
since `warning` itself has such a wrapper.