API for setting animation timeouts
This commit provides an API for setting timout-delay lengths for each
kind of animation in ReactCSSTransitionGroup.  If an animation hasn't
finished within the specified time (either because the browser isn't
sending the transitionend event or because the animation is
exceptionally long) it will be cleaned-up.  Specifying these timeouts is
optional; they default to 5 seconds to preserve existing behaviour.
Any animations which take longer than 5 seconds will be broken by this
change, but that should be rare and -- using this new API -- developers
can specify a longer timeout for such animations.

This whole system isn't ideal (since developers need to write the
animation duration in CSS as-well-as in the component), but this
approach allows developers to produce apparently reliable animations
(which is currently impossible using ReactCSSTransitionGroup) by
specifying the exact expected animation duration.