Optional transition clean-up timeouts
Transition cleanup on a timeout is now wholly opt-in, making the
augmented ReactCSSTransitionGroup API fully backwards compatible. Now,
the transition cleanup timeout is set-up only if a timeout-delay was
specified by the caller.

The dev-mode 'no-transitionend' warning has been reverted to a 5000ms
timeout and only applies if no timeout was specified by the caller.
Previously, this warning would trigger on every transition if the exact
duration of the animation was specified (because the timeout raced with
the transitionend event and always won).

This essentially creates two usage modes for the component: a more
automated default mode (which inherits the unreliability of the
transitionend events that it listens to); and an advanced mode where the
animation duration is specified explicitly and all cleanup is performed
with timeouts. It is also possible to specify over-long timeouts as
fail-safes and use the transitionend event where possible. All three of
these usage patterns are now discussed in the documentation (where users
will hopefully find themselves if they're having reliability problems
with the default mode).